app/requirements.txt
numpy==1.22.1
Pillow==9.2.0
streamlit==1.11.0
